Make process executions non-blocking

+++ backend/src/main/java/at/uibk/dps/rm/service/deployment/DeploymentService.java
@@ -19,7 +19,7 @@ public interface DeploymentService {
         return new DeploymentServiceVertxEBProxy(vertx, address);
     }
 
-    Future<Long> deploy(JsonObject data);
+    Future<Integer> deploy(JsonObject data);
 
     Future<Void> terminate(long resourceId);
 }
